consolidate share extension view code
